==> scripts/00-install_packages.R <==
#### Preamble ####
# Purpose: Installs the necessary packages. 
# Author: Rahma Binth Mohammad
# Date: July 20, 2024
# Contact: rahma.binthmohammad@mail.utoronto.ca
# License: MIT

#### Download packages ####
install.packages("tidyverse")
install.packages("cansim")
install.packages("dplyr")
install.packages("rstanarm")
